requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
gTTS==2.3.2
Werkzeug==2.3.7
python-multipart==0.0.6
Pillow==10.0.1